refactor: move LogFetcher to RunnerBarCore/Services (#926)
- Delete Sources/RunnerBar/Services/LogFetcher.swift
- Add Sources/RunnerBarCore/Services/LogFetcher.swift with public API
- Remove import RunnerBarCore (can't import yourself)
- ghRaw() and ghBinaryPath() already in Core — resolve directly
- No import changes needed in callers (covered by @_exported in Exports.swift)
